pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bitbound"
version = "1.0.0"
authors = [
    { name = "BitBound Team", email = "contact@bitbound.io" }
]
description = "High-Level Embedded Python Library for declarative hardware programming"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.7"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: Implementation :: MicroPython",
    "Topic :: System :: Hardware",
    "Topic :: Software Development :: Embedded Systems",
    "Topic :: Scientific/Engineering",
]
keywords = ["embedded", "micropython", "hardware", "iot", "sensors", "gpio", "i2c", "spi", "raspberry-pi", "esp32", "arduino", "mqtt", "ota", "wifi"]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "pytest-asyncio>=0.21",
    "black>=22.0",
    "mypy>=0.9",
    "build",
    "twine",
]
network = [
    "paho-mqtt>=1.6",
    "websocket-client>=1.0",
    "requests>=2.28",
]
all = [
    "paho-mqtt>=1.6",
    "websocket-client>=1.0",
    "requests>=2.28",
    "pyserial>=3.5",
]

[project.scripts]
bitbound = "bitbound.cli:main"

[project.urls]
Homepage = "https://github.com/bitbound/bitbound"
Documentation = "https://bitbound.readthedocs.io/"
Repository = "https://github.com/bitbound/bitbound"
Issues = "https://github.com/bitbound/bitbound/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["bitbound*"]

[tool.black]
line-length = 100
target-version = ['py37', 'py38', 'py39', 'py310', 'py311', 'py312']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
